Create Data folder for all Get and Create operations Autofac registering Move stuff out of CheckoutManager
