+++ ReactShop.Core/DTOs/CartDTO.cs
@@ -4,10 +4,6 @@ namespace ReactShop.Core.DTOs
 {
     public class CartDTO
     {
-        public CartDTO()
-        {
-        }
-
         public decimal Subtotal { get; set; }
         public decimal DiscountRate { get; set; }
         public decimal DiscountValue { get; set; }
