+++ ReactShop.Core/Data/Cart/ISaveCart.cs
@@ -1,0 +1,9 @@
+﻿using ReactShop.Core.DTOs;
+
+namespace ReactShop.Core.Data.Cart
+{
+    public interface ISaveCart
+    {
+        void Save(CartDTO cart);
+    }
+}
