+++ ReactShop.Core/Data/Cart/SaveCart.cs
@@ -1,0 +1,12 @@
+﻿using ReactShop.Core.DTOs;
+
+namespace ReactShop.Core.Data.Cart
+{
+    public class SaveCart : ISaveCart
+    {
+        public void Save(CartDTO cart)
+        {
+            throw new System.NotImplementedException();
+        }
+    }
+}
