+++ ReactShop.Core/Data/Orders/GetOrders.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+using ReactShop.Core.Entities;
+
+namespace ReactShop.Core.Data.Orders
+{
+    public class GetOrders : IGetOrders
+    {
+        public IEnumerable<Order> Get()
+        {
+            throw new System.NotImplementedException();
+        }
+    }
+}
