+++ ReactShop.Core/Data/Orders/ICreateOrder.cs
@@ -1,0 +1,9 @@
+﻿using ReactShop.Core.DTOs;
+
+namespace ReactShop.Core.Data.Orders
+{
+    public interface ICreateOrder
+    {
+        int Create(CartDTO cart);
+    }
+}
