+++ ReactShop.Core/Data/Orders/IGetOrders.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using ReactShop.Core.Entities;
+
+namespace ReactShop.Core.Data.Orders
+{
+    public interface IGetOrders
+    {
+        IEnumerable<Order> Get();
+    }
+}
