+++ ReactShop.Core/Data/Products/IGetProducts.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using ReactShop.Core.DTOs;
+
+namespace ReactShop.Core.Data.Products
+{
+    public interface IGetProducts
+    {
+        IEnumerable<ProductDTO> Get();
+        ProductDTO GetBySku(string sku);
+    }
+}
