+++ ReactShop.Core/IInitializeDB.cs
@@ -1,0 +1,7 @@
+﻿namespace ReactShop.Core
+{
+    public interface IInitializeDB
+    {
+        Context Init();
+    }
+}
